created close button and branch icon components
